Add files via upload
This is a template for a sample bio webpage.  It includes fun, interactive animations, and it primarily uses jQuery in the javascript file.  The image can be substituted for any other.  
some features include: 
- navigation menu
- interactive resume compenents
- formatted for different screens

etc.

I will still be making some tweaks, but until I get my website up and running, this is a great way to market my abilities.  Feel free to use the template!